audio_to_image_ai/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
langchain-community
langchain-ollama
diffusers
transformers
torch
accelerate
pillow
soundfile
numpy
